Long int for handling sizes
git-svn-id: file:///home/roberto/Git/svn@1225 1a721e8d-d38c-4e91-ab98-6a0be4d754af

+++ trunk/Xmipp/Lib/XmippData/xmippMatrices1D.hh
@@ -343,7 +343,7 @@ public:
        if (__m==NULL) REPORT_ERROR(1001,"Resize: no memory left");
        __dim=XSIZE(*this)=dim;
        __spcdim=1;
-       for (int i=0; i<__dim; i++) __m[i]=0;
+       for (long int i=0; i<__dim; i++) __m[i]=0;
    }
 
    /** Copy constructor.
